B: specify group name for order-instanced-pipelines
rather than needing to specify the pipeline name multiple times:

fly oip -p group/k1:v1 -p group/k2:v2

...you now specify the group name once and the list of instance vars
separately:

fly oip -g group -p k1:v1 -p k2:v2
