+++ fly/commands/internal/flaghelpers/resource_flag_test.go
@@ -72,7 +72,7 @@ var _ = Describe("ResourceFlag", func() {
 		{
 			desc: "malformed instance var",
 			flag: "some-pipeline/branch=master/some-resource",
-			err:  "argument format should be <pipeline>/<key:value>/<resource>",
+			err:  "instance vars should be formatted as <key1:value1>(,<key2:value2>)",
 		},
 	} {
 		tt := tt
