+++ go-concourse/concourse/pipelines.go
@@ -3,7 +3,6 @@ package concourse
 import (
 	"bytes"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 
@@ -58,23 +57,10 @@ func (team *team) OrderingPipelines(pipelineNames []string) error {
 	}, &internal.Response{})
 }
 
-func (team *team) OrderingPipelinesWithinGroup(pipelineRefs []atc.PipelineRef) error {
-	if len(pipelineRefs) == 0 {
-		return errors.New("no pipelines given")
-	}
-	pipelineName := pipelineRefs[0].Name
-	var instanceVars []atc.InstanceVars
-
-	for _, pipelineRef := range pipelineRefs {
-		if pipelineRef.Name != pipelineName {
-			return fmt.Errorf("All given instance pipelines must belong to the same pipeline: %s did not match %s", pipelineRef.Name, pipelineName)
-		}
-		instanceVars = append(instanceVars, pipelineRef.InstanceVars)
-	}
-
+func (team *team) OrderingPipelinesWithinGroup(groupName string, instanceVars []atc.InstanceVars) error {
 	params := rata.Params{
 		"team_name":     team.Name(),
-		"pipeline_name": pipelineName,
+		"pipeline_name": groupName,
 	}
 
 	buffer := &bytes.Buffer{}
