+++ go-concourse/concourse/team.go
@@ -71,7 +71,7 @@ type Team interface {
 	CreateBuild(plan atc.Plan) (atc.Build, error)
 	Builds(page Page) ([]atc.Build, Pagination, error)
 	OrderingPipelines(pipelineNames []string) error
-	OrderingPipelinesWithinGroup(pipelineRefs []atc.PipelineRef) error
+	OrderingPipelinesWithinGroup(groupName string, instanceVars []atc.InstanceVars) error
 
 	CreateArtifact(io.Reader, string, []string) (atc.WorkerArtifact, error)
 	GetArtifact(int) (io.ReadCloser, error)
